+++ minijinja/src/vm/closure_object.rs
@@ -4,12 +4,42 @@ use std::sync::{Arc, Mutex};
 
 use crate::value::{Object, ObjectKind, StructObject, Value};
 
+/// Closure cycle breaker utility.
+///
+/// The closure tracker is a crude way to forcefully break the cycles
+/// caused by closures on teardown of the state.  Whenever a closure is
+/// exposed to the engine it's also passed to the [`track_closure`] function.
+#[derive(Default)]
+pub struct ClosureTracker {
+    closures: Mutex<BTreeMap<usize, Arc<Closure>>>,
+}
+
+impl ClosureTracker {
+    /// This accepts a closure as value and registers it in the
+    /// tracker for cycle breaking.
+    ///
+    /// This only registers each closure once.
+    pub(crate) fn track_closure(&self, closure: Arc<Closure>) {
+        let ptr = &closure as &Closure;
+        let id = ptr as *const _ as usize;
+        self.closures.lock().unwrap().insert(id, closure);
+    }
+}
+
+impl Drop for ClosureTracker {
+    fn drop(&mut self) {
+        for closure in self.closures.lock().unwrap().values() {
+            closure.clear();
+        }
+    }
+}
+
 /// Utility to enclose values for macros.
 ///
 /// See `closure` on the [`Frame`] for how it's used.
 #[derive(Debug, Default)]
 pub(crate) struct Closure {
-    values: Mutex<BTreeMap<Arc<str>, Value>>,
+    values: Arc<Mutex<BTreeMap<Arc<str>, Value>>>,
 }
 
 impl Closure {
@@ -25,6 +55,13 @@ impl Closure {
         if !values.contains_key(key) {
             values.insert(Arc::from(key), f());
         }
+    }
+
+    /// Clears the closure.
+    ///
+    /// This is required to break cycles.
+    pub fn clear(&self) {
+        self.values.lock().unwrap().clear();
     }
 }