+++ minijinja/src/vm/context.rs
@@ -6,8 +6,10 @@ use std::sync::Arc;
 use crate::environment::Environment;
 use crate::error::{Error, ErrorKind};
 use crate::value::{OwnedValueIterator, Value, ValueRepr};
-use crate::vm::closure_object::Closure;
 use crate::vm::loop_object::Loop;
+
+#[cfg(feature = "macros")]
+use crate::vm::closure_object::Closure;
 
 type Locals<'env> = BTreeMap<&'env str, Value>;
 
@@ -32,6 +34,7 @@ pub(crate) struct Frame<'env> {
     // duplicated into the closure.  Macros declared on that level, then share
     // the closure object to enclose the parent values.  This emulates the
     // behavior of closures in Jinja2.
+    #[cfg(feature = "macros")]
     pub(crate) closure: Option<Arc<Closure>>,
 }
 
@@ -48,6 +51,7 @@ impl<'env> Frame<'env> {
             locals: Locals::new(),
             ctx,
             current_loop: None,
+            #[cfg(feature = "macros")]
             closure: None,
         }
     }
@@ -199,8 +203,11 @@ impl<'env> Context<'env> {
     /// Stores a variable in the context.
     pub fn store(&mut self, key: &'env str, value: Value) {
         let top = self.stack.last_mut().unwrap();
-        if let Some(ref closure) = top.closure {
-            closure.store(key, value.clone());
+        #[cfg(feature = "macros")]
+        {
+            if let Some(ref closure) = top.closure {
+                closure.store(key, value.clone());
+            }
         }
         top.locals.insert(key, value);
     }
@@ -234,13 +241,13 @@ impl<'env> Context<'env> {
     /// This means that if you override a variable referenced by a macro after
     /// including in the parent template, it will not override the value seen by
     /// the macro.
-    #[cfg(feature = "multi_template")]
+    #[cfg(all(feature = "multi_template", feature = "macros"))]
     pub fn take_closure(&mut self) -> Option<Arc<Closure>> {
         self.stack.last_mut().unwrap().closure.take()
     }
 
     /// Puts the closure back.
-    #[cfg(feature = "multi_template")]
+    #[cfg(all(feature = "multi_template", feature = "macros"))]
     pub fn reset_closure(&mut self, closure: Option<Arc<Closure>>) {
         self.stack.last_mut().unwrap().closure = closure;
     }