+++ minijinja/src/defaults.rs
@@ -75,6 +75,8 @@ pub(crate) fn get_builtin_filters() -> BTreeMap<Cow<'static, str>, filters::Boxe
         rv.insert("default".into(), BoxedFilter::new(filters::default));
         rv.insert("round".into(), BoxedFilter::new(filters::round));
         rv.insert("abs".into(), BoxedFilter::new(filters::abs));
+        rv.insert("int".into(), BoxedFilter::new(filters::int));
+        rv.insert("float".into(), BoxedFilter::new(filters::float));
         rv.insert("attr".into(), BoxedFilter::new(filters::attr));
         rv.insert("first".into(), BoxedFilter::new(filters::first));
         rv.insert("last".into(), BoxedFilter::new(filters::last));