+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to MiniJinja are documented here.
 
 ## 1.0.10
 
-- Added `int` and `float` filters.
+- Added `int` and `float` filters.  #372
 
 ## 1.0.9